refactor(iree): set CPU stack allocation limit as an LLVM target option (#34)
In order to systemically address `exceeded stack allocation limit` failures for the CPU pipeline, we first need a way of conveying the limit itself to the whole pipeline. Therefore, create a new context-wide attribute, as well as a bound CLI option to replace the one specific to `LLVMCPUCheckIRBeforeLLVMConversionPass`.

A further direction for improvement would be unit-testing all the target properties & corresponding CLI options on the basis of this PR.