feat(CUDA): add new target for testing (#30)
We want to try a simplified GPU pipeline for CUDA and test on top of it

This adds a new `cellar-cuda` target backend that implements a simplified CUDA pipeline. The key changes include:

- A new CUDA CodeGen library that will be used to house our new pipeline for CUDA. 
- CUDAKernelConfig contains all the current strategies for lowering down to PTX. We will have for now TileAndFuse, VectorDistribute and GPUDistribute as our main three strategies for now.  

Most of the code is copied from the original target. I would say review and check how everything is structured rather than looking at the details themselves.