feat(iree): fuse elementwise generic ops into `iree_linalg_ext.scan` ops
Enhance `iree_linalg_ext.scan` support in the pipeline by introducing a basic
fusion pattern for generic ops that are determined as unary elementwise
operations. Provided that the full shapes match between the GenericOp
producer and the ScanOp consumer, we can simply repeat the unary operation
directly during the input data indexation of a ScanOp reduction routine.

This is essentially serving as a workaround for LLVMCPU pipeline, where
stack-bound allocations tend to exceed the size limit in the presence of a
ScanOp that just receives the pipeline's default tiling configuration.
Further enhancements should include tiling config fine-tuning for LinalgExt
operations, concious restrictions of work-group level tiling depending
on the predicted size of stack-bound allocas within a dispatch. However,
the fusion itself is beneficial, and we should look to support more producer
cases, expanding the logic onto other pre-defined LinalgExt ops.