feat(CellarCuda): add pipeline codgen enum (#52)
In order to add a completely new pipeline, we are required to add a new enum to the list of available codegen pipelines.

Up till now we have been using the pre-existing pipelines, but since we want to use our own custom codegen path we are now required to add it here.  

This most likely may have to happen for each custom target we add...