fix(CloneProducersIntoWorkgroups): DPS Style Op Check
The issue that triggers the failure is:

```
module {
  util.func public @main(%arg0: tensor<1x1xf32>) -> tensor<1x9xf32> {
    %cst = arith.constant dense<[[1.000000e+00, 2.000000e+00, 3.000000e+00, 4.000000e+00, 5.000000e+00, 6.000000e+00, 7.000000e+00, 8.000000e+00, 9.000000e+00]]> : tensor<1x9xf32>
    %cst_0 = arith.constant dense<[[1.000000e+00, 1.000000e+00, 2.000000e+00, 2.000000e+00, 3.000000e+00, 3.000000e+00, 4.000000e+00, 4.000000e+00, 5.000000e+00]]> : tensor<1x9xf32>
    %0 = linalg.matmul ins(%arg0, %cst : tensor<1x1xf32>, tensor<1x9xf32>) outs(%cst_0 : tensor<1x9xf32>) -> tensor<1x9xf32>
    util.return %0 : tensor<1x9xf32>
  }
}
```

Because of the size, we generate two mmt4d kernel calls that get wrapped
in a workgroup. Because of this the workgroup verification check will
find the ForAllOp that contains the MMT4D operation and check if there
are any global loads.

Because the constant was cloned into the workgroup and it is on the
output of the linalg operation, it was marked with the storage_buffer
attribute and so triggers the failure.

The following change prevents output tensors from being cloned into the
workgroup even if they are quite small as they are deemed writable and
by definition of Destinantion Style Passing are an output of the
function not an input so it should not be optimized into the loop.