feat(codegen): use control function to prevent fusion-inhibiting folding patterns (#69)
This is step 2/3 of fixing the large vector sizes found compilation failure (roo-389). This PR adds the control function that would allow us to prevent folding pack/unpack operations into (transpose) ops that has multiple uses, e.g. the result is stored into a dispatch. 

To begin with, ending up with such dispatches is also problematic in means of memory footprint, but I guess fixing that would be a much more involved issue. So, this PR enables fusion by blocking inhibiting folding patterns. See IREE issue #20896 for more details.

The change I made, and therefore the test that would be changed is actually target-agnostic. Although generally, the pass that utilizes it isn't. So I just put it into one of them. If anyone has a better idea of where to put the test, please tell me :) Theoretically, I can put it in a new file that is dedicated to checking if these patterns are blocked or not - also would be a place to put future patterns to block over there. But I assume this would rarely be updated, so idk.

TODO:
- [x] Add and/or adjust tests.